Rename file and modify imports
