+++ scrapy/signalmanager.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING, Any, List, Tuple
 
 from pydispatch import dispatcher
 
-from scrapy.utils import signal as _signal
+from scrapy.utils import signal_ops as _signal
 
 if TYPE_CHECKING:
     from twisted.internet.defer import Deferred
