+++ scrapy/utils/signal_ops.py
