+++ tests/test_engine.py
@@ -37,7 +37,7 @@ from scrapy.item import Field, Item
 from scrapy.linkextractors import LinkExtractor
 from scrapy.signals import request_scheduled
 from scrapy.spiders import Spider
-from scrapy.utils.signal import disconnect_all
+from scrapy.utils.signal_ops import disconnect_all
 from scrapy.utils.test import get_crawler
 from tests import get_testdata, tests_datadir
